// src/generated/models/Cgmp_changetemplatesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cgmp_changetemplatesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cgmp_changetemplatesstatecode = keyof typeof Cgmp_changetemplatesstatecode;
export const Cgmp_changetemplatesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cgmp_changetemplatesstatuscode = keyof typeof Cgmp_changetemplatesstatuscode;

export interface Cgmp_changetemplatesBase {
  cgmp_category?: string;
  cgmp_changetemplateid: string;
  cgmp_changetype?: string;
  cgmp_country?: string;
  cgmp_description?: string;
  cgmp_impactlevel?: number;
  cgmp_isemergency?: boolean;
  cgmp_location?: string;
  cgmp_name?: string;
  cgmp_projectids?: string;
  cgmp_region?: string;
  cgmp_risklevel?: number;
  cgmp_timeline?: string;
  cgmp_uatrequired?: boolean;
  cgmp_uatusers?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cgmp_changetemplatesstatecode;
  statuscode?: Cgmp_changetemplatesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cgmp_changetemplates extends Cgmp_changetemplatesBase {
  cgmp_isemergencyname?: string;
  cgmp_uatrequiredname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}